pyproject.toml
[project]
name = "k8s-cli"
version = "1.0.0"
description = "SkyPilot-compatible Kubernetes task launcher"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn>=0.32.0,<1.0.0",
    "kr8s>=0.17.0,<1.0.0",
    "pydantic>=2.9.0,<3.0.0",
    "pyyaml>=6.0.0,<7.0.0",
    "typer>=0.12.0,<1.0.0",
    "httpx>=0.27.0,<1.0.0",
    "rich>=13.0.0,<14.0.0",
]

[project.scripts]
k8s-cli = "k8s_cli.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
